+++ tests/core/source_methods/test_count.py
@@ -1,3 +1,5 @@
+import pytest
+
 import onetick.py as otp
 
 
@@ -11,5 +13,6 @@ def test_count(m_session):
     data = otp.merge([otp.Ticks(X=[1])], symbols=otp.Ticks(SYMBOL_NAME=['A', 'B']))
     assert data.count() == 2
 
-    data = otp.merge([otp.Ticks(X=[1])], symbols=otp.Empty())
-    assert data.count() == 0
+    with pytest.warns(UserWarning, match='Eval statement returned no symbols'):
+        data = otp.merge([otp.Ticks(X=[1])], symbols=otp.Empty())
+        assert data.count() == 0